+++ public/data/mockFeeds.json
@@ -1,0 +1,23 @@
+[
+  {
+    "id": 1,
+    "userName": "wecode",
+    "content": " Welcome to world best coding bootcamp!",
+
+    "src": "./images/jaehyunLee/Main/iu.jpg"
+  },
+  {
+    "id": 2,
+    "userName": "dlwlrma",
+    "content": " Welcome to world best coding bootcamp!",
+
+    "src": "./images/jaehyunLee/Main/iu2.jpg"
+  },
+  {
+    "id": 3,
+    "userName": "wecode",
+    "content": " Welcome to world best coding bootcamp!",
+
+    "src": "./images/jaehyunLee/Main/iu3.jpg"
+  }
+]
