+++ public/data/mockReplies.json
@@ -1,0 +1,20 @@
+[
+  {
+    "id": 1,
+    "userName": "wecode",
+    "content": " Welcome to world best coding bootcamp!",
+    "isLike": true
+  },
+  {
+    "id": 2,
+    "userName": "joonsikyang",
+    "content": " Hi there.",
+    "isLike": false
+  },
+  {
+    "id": 3,
+    "userName": "jayPark",
+    "content": " Hey.",
+    "isLike": false
+  }
+]
